Code cleanup

+++ Api/0.0.0.1/EtAlii.Ubigia.Api.Functional.Scripting.NET47.SignalR.Tests/1. Logical/UnitTests/Win32ContentManager.Tests.cs
@@ -25,12 +25,9 @@
         protected virtual void Dispose(bool disposing)
         {
             // Cleanup
-            if (disposing)
+            if (disposing && File.Exists(_testImageFileName))
             {
-                if (File.Exists(_testImageFileName))
-                {
-                    File.Delete(_testImageFileName);
-                }
+                File.Delete(_testImageFileName);
             }
         }
